pyproject.toml
[project]
name = "basic-dbt-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagster>=1.9.10",
    "dagster-dbt>=0.25.10",
    "dagster-webserver>=1.9.10",
    "dbt-core>=1.9.1",
    "dbt-duckdb>=1.9.1",
    "pandas>=2.2.3",
]
